# By Igor Solovey (22) and others
# Via Igor Solovey (22) and others
* commit 'db92d55fb0e09551257f8f2fd79d8516e56aa00a [formerly 3da43286117094ff5df532eadb1007037b3559fe]': (123 commits)
  DF-9778 review fix
  [DF-9778] Fixed wrong cast
  DF-9778 win10 fix
  DF-9778 review fix
  DF-9778 review fix
  DF-9778 review fix
  DF-9778 review fix
  DF-9778 win10 fix
  DF-9778 win10 fix
  DF-9778 win10 fix, Werror
  eliminated some warnings in GL-CommandBuffer (MacOS);
  DF-9778 fix error on windows
  DF-9778 fix warnigns
  [DF-9778] libuv rebuild with 10.8 deployment target
  DF-9778
  DF-9778 fix warnigns
  [DF-9778] VertexBufferGLES2 warning fix
  [DF-9778] rhi_IndexBufferGLES2 warnings fix.
  DF-9778 warning fix
  [DF-9778] ProgGLES2 warning fix
  ...


Former-commit-id: 71cb26bf9f03982f58c5731d7533c12d1be72364

+++ Sources/Internal/Platform/TemplateWin32/DeviceInfoWin32.cpp
@@ -201,7 +201,7 @@ String DeviceInfoPrivate::GetUDID()
     MD5::ForData(reinterpret_cast<const uint8*>(res.c_str()), static_cast<uint32>(res.size()), md5Digest);
 
     String digest(MD5::MD5Digest::DIGEST_SIZE * 2 + 1, '\0');
-    MD5::HashToChar(md5Digest, const_cast<char8*>(digest.data()), digest.size());
+    MD5::HashToChar(md5Digest, const_cast<char8*>(digest.data()), static_cast<uint32>(digest.size()));
     return digest;
 }
